command/cat: move cat logic to cp
